% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_sn.R
\name{cusum_sn}
\alias{cusum_sn}
\title{CUSUM Control Chart for Grouped Data (Based on Signs)}
\usage{
cusum_sn(
  X,
  med = 0,
  k = 0.5,
  h = 5,
  group_by_col = F,
  plot = T,
  side = "two.sided"
)
}
\arguments{
\item{X}{A numeric matrix or data frame containing the data.}

\item{med}{The median value to use as the reference.
Default is 0.}

\item{k}{The reference value for the change detection.
Default is 0.5.}

\item{h}{The control limit value.
Default is 5.}

\item{group_by_col}{A logical value indicating whether to treat
columns as separate groups. Default is FALSE.}

\item{plot}{A logical value indicating whether to plot the control chart.
Default is TRUE.}

\item{side}{The side of the control chart,
can be "two.sided", "lower", or "upper".}
}
\value{
If `plot` is set to TRUE,
the function will generate a control chart plot.
If `plot` is set to FALSE,
the function will return the first out-of-control point (index)
based on the specified `side`.
}
\description{
Calculates and plots the CUSUM Control Chart for grouped data based on signs.
}
\details{
The function calculates the CUSUM Control Chart 
for grouped data based on signs. It calculates the number of observations 
above and below the median for each group, and then applies the CUSUM 
algorithm to detect changes in the process.
The control limits are determined by the value 
of `h` and the `side` parameter.
The function can plot the control chart 
using the `plot_chart` function or return the index 
of the first out-of-control observation.
}
\examples{
# Generate example data
set.seed(42)
X <- matrix(rnorm(100), nrow = 20)

# Calculate CUSUM Control Chart for grouped data
cusum_sn(X, med = 0, k = 0.5, h = 5, group_by_col = FALSE, 
plot = TRUE, side = 'two.sided')

}
\seealso{
\code{\link{plot_chart}}
}
