% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_sn_arl_calc.R
\name{cusum_sn_arl_calc}
\alias{cusum_sn_arl_calc}
\title{Average Run Length (ARL) Calculation for CUSUM Control Chart}
\usage{
cusum_sn_arl_calc(
  n = 10,
  k = 0.5,
  h = 3,
  t = 10000,
  n_sim = 10000,
  group_by_col = F,
  side = "two.sided"
)
}
\arguments{
\item{n}{The number of observations in each group. Default is 10.}

\item{k}{The reference value for the change detection. Default is 0.5.}

\item{h}{The control limit value. Default is 3.}

\item{t}{The number of samples (lower values are problematic).}

\item{n_sim}{The number of simulations to perform.}

\item{side}{The side of the control chart, can be "two.sided",
"lower", or "upper".}
}
\value{
The average run length (ARL) for the CUSUM Control Chart.
}
\description{
Calculates the average run length (ARL) for the CUSUM Control Chart.
}
\details{
The function calculates the average run length (ARL) for
  the CUSUM Control Chart.
  It performs multiple simulations based on the specified
  parameters and returns the average ARL.
  Each simulation consists of generating random data with
  the given number of observations (`n`)
  and calculating the CUSUM Control Chart using the `cusum_sn` function.
  The ARL is the average number of observations until the control
  chart signals an out-of-control condition.
  If NA values occur during the simulations, a warning message
  will be displayed, and those NA values will be excluded from the
  calculation.
}
\examples{
# Calculate ARL for CUSUM Control Chart
cusum_sn_arl_calc(n = 1, k = 0.5, h = 3, t = 10000, n_sim = 10000,
 group_by_col = FALSE, side = 'two.sided')

}
