% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_sr.R
\name{cusum_sr}
\alias{cusum_sr}
\title{CUSUM Control Chart for Grouped Data (Based on Rank Statistics)}
\usage{
cusum_sr(
  X,
  med = 0,
  k = 0.5,
  h = 5,
  group_by_col = F,
  plot = T,
  side = "two.sided"
)
}
\arguments{
\item{X}{A numeric matrix or data frame containing the data.}

\item{med}{The median value to use as the reference. Default is 0.}

\item{k}{The reference value for the change detection. Default is 0.5.}

\item{h}{The control limit value. Default is 5.}

\item{group_by_col}{A logical value indicating whether to treat
columns as groups. If TRUE, the matrix will be transposed. Default is FALSE.}

\item{plot}{A logical value indicating whether to plot the control chart.
Default is TRUE.}

\item{side}{The side of the chart to plot.
Must be one of 'two.sided', 'lower', or 'upper'. Default is 'two.sided'.}
}
\value{
If `plot` is set to FALSE, the function returns
the index of the first out-of-control observation.
If `plot` is set to TRUE, the function generates a control chart plot.
}
\description{
Calculates and plots the CUSUM Control Chart for grouped data based
on rank statistics.
}
\details{
The function calculates the CUSUM Control Chart for
grouped data based on rank statistics.
It calculates the rank statistic for each group using the Wilcoxon
rank sum test, and then applies the CUSUM algorithm to detect
changes in the process. The control limits are determined by
the value of `h` and the `side` parameter.
The function can plot the control chart using the `plot_chart`
function or return the index of the first out-of-control observation.
}
\examples{
# Generate example data
set.seed(42)
X <- matrix(rnorm(100), nrow = 20)

# Calculate CUSUM Control Chart for grouped data
cusum_sr(X, med = 0, k = 0.5, h = 5, group_by_col = FALSE,
         plot = TRUE, side = 'two.sided')

}
\seealso{
\code{\link{plot_chart}}
}
