% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_sr_arl_calc.R
\name{cusum_sr_arl_calc}
\alias{cusum_sr_arl_calc}
\title{Calculate Average Run Length (ARL) for CUSUM Control Chart (Based on Signed-ranks)}
\usage{
cusum_sr_arl_calc(
  n = 10,
  k = 0.5,
  h = 3,
  t = 10000,
  n_sim = 10000,
  group_by_col = FALSE,
  side = "two.sided"
)
}
\arguments{
\item{n}{The sample size. Default is 10.}

\item{k}{The reference value for the CUSUM chart. Default is 0.5.}

\item{h}{The decision interval for the CUSUM chart. Default is 3.}

\item{t}{The number of observations per sample. Default is 10000.}

\item{n_sim}{The number of simulated samples. Default is 10000.}

\item{group_by_col}{A logical value indicating whether the samples 
are grouped by columns.
  If TRUE, each column represents a sample. If FALSE (default), 
  each row represents a sample.}

\item{side}{The side of the control chart. Must be one of 'two.sided', 
'lower', or 'upper'.
  Default is 'two.sided'.}
}
\value{
The average run length (ARL) of the CUSUM control chart.
}
\description{
This function calculates the Average Run Length (ARL) for a 
CUSUM control chart
based on signed-ranks. It generates random samples and applies the 
CUSUM control
chart procedure to each sample to calculate the ARL.
}
\details{
The function generates \code{n_sim} random samples of size \code{n} 
and applies
the CUSUM control chart procedure to each sample. 
The CUSUM control chart is based on
signed-ranks and uses the reference value \code{k} and the 
decision interval \code{h}.
The ARL is calculated as the average number of observations until 
the CUSUM chart signals
out-of-control.
}
\examples{
arl <- cusum_sr_arl_calc(n = 10, k = 0.5, h = 3, t = 10000, 
                        n_sim = 10000, side = 'two.sided')
arl

}
\seealso{
\code{\link{cusum_sr}} for the CUSUM control chart procedure 
based on signed-ranks.
}
