% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\docType{data}
\name{data1}
\alias{data1}
\title{CuSum-SR Data from Table 4.7a (Chakraborti and Graham, 2019)}
\format{
An example of a data matrix for the CuSum-SR
}
\usage{
data1
}
\description{
The data was extracted from the book of Chakraborti and Graham (2019)
that contains 20 groups with 4 observations each
}
\keyword{datasets}
