% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma_sn.R
\name{ewma_sn}
\alias{ewma_sn}
\title{EWMA SN Control Chart for Grouped Data}
\usage{
ewma_sn(
  X,
  med = 0,
  lambda = 0.05,
  L = 2.667,
  group_by_col = FALSE,
  side = "two.sided"
)

ewma_sn(
  X,
  med = 0,
  lambda = 0.05,
  L = 2.667,
  group_by_col = FALSE,
  side = "two.sided"
)
}
\arguments{
\item{X}{A numeric matrix or data frame containing the data.}

\item{med}{The median value to use as the reference. Default is 0.}

\item{lambda}{The decay factor for the EWMA. Default is 0.05.}

\item{L}{The multiplier for the control limits. Default is 2.667.}

\item{group_by_col}{A logical value indicating whether to treat columns as groups.
If TRUE, the matrix will be transposed. Default is FALSE.}

\item{side}{The side of the chart to plot. Must be one of 'two.sided', 'lower', or 'upper'.
Default is 'two.sided'.}

\item{plot}{A logical value indicating whether to plot the control chart.
Default is TRUE.}
}
\value{
If `plot` is set to TRUE,
the function will generate a control chart plot.
If `plot` is set to FALSE,
the function will return the first out-of-control point (index)
based on the specified `side`.

The function generates a control chart plot and returns the average run length (ARL).
}
\description{
This function calculates the Exponentially Weighted Moving Average
(EWMA) control chart based on the signs (SN) statistic for grouped data.

Calculates the EWMA for grouped data based on signs.
}
\details{
The function calculates the EWMA based on signs for grouped data.
  It calculates the number of observations above and below the median for each group,
  and then applies the EWMA formula to obtain the control statistics.
  The control limits are calculated based on the side parameter.
  The function also plots the control chart using the plot_chart function
  and calculates the average run length (ARL) based on simulated data.
}
\examples{
# Generate random grouped data
X <- matrix(rnorm(100), nrow = 10)
# Create the EWMA SN control chart
ewma_sn(X)

# Generate example data
set.seed(42)
X <- matrix(rnorm(100), nrow = 10)

# Calculate EWMA for grouped data
ewma_sn(X, med = 0, lambda = 0.05, L = 2.667, group_by_col = FALSE, side = 'two.sided')

}
\seealso{
\code{\link{ewma_sn_simulation}}, \code{\link{plot_chart}}
}
