% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma_sr.R
\name{ewma_sr}
\alias{ewma_sr}
\title{EWMA Control Chart Based on Signed-Ranks Statistic}
\usage{
ewma_sr(
  X,
  med = 0,
  lambda = 0.05,
  L = 2.472,
  group_by_col = FALSE,
  plot = TRUE,
  side = "two.sided"
)
}
\arguments{
\item{X}{A matrix or data frame of observations.}

\item{med}{The center value for the control chart. Default is 0.}

\item{lambda}{The smoothing parameter for the EWMA calculation. 
Default is 0.05.}

\item{L}{The control limit multiplier for the UCL and LCL calculation.}

\item{group_by_col}{A logical value indicating whether to treat 
columns as separate groups. Default is FALSE.}

\item{plot}{A logical value indicating whether to plot the control chart. 
Default is TRUE.}

\item{side}{The side of the control chart, 
can be "two.sided", "lower", or "upper".}
}
\value{
If `plot` is set to TRUE, 
the function will generate a control chart plot.
If `plot` is set to FALSE, 
the function will return the first out-of-control point (index) 
based on the specified `side`.
}
\description{
This function constructs and plots an EWMA Control Chart 
based on the signed-ranks statistic.
}
\examples{
# Create and plot an EWMA Control Chart based on signed-ranks statistic
data <- matrix(rnorm(100), ncol = 5)
ewma_sr(data, med = 0, lambda = 0.05, L = 2.472, 
group_by_col = FALSE, plot = TRUE, side = "two.sided")

}
