% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma_mw.R, R/ewma_wr.R
\name{ewma_wr}
\alias{ewma_wr}
\title{EWMA Control Chart based on the Wilcoxon Rank Sum Test}
\usage{
ewma_wr(
  X,
  Y,
  lambda = 0.05,
  L = 2.447,
  group_by_col = F,
  plot = T,
  side = "two.sided"
)

ewma_wr(
  X,
  Y,
  lambda = 0.05,
  L = 2.447,
  group_by_col = F,
  plot = T,
  side = "two.sided"
)
}
\arguments{
\item{X}{The reference value.}

\item{Y}{The data matrix or vector to be compared.}

\item{lambda}{The weight parameter for the EWMA calculation
(default is 0.05).}

\item{L}{The control limit multiplier
(default is 2.447).}

\item{group_by_col}{A logical value indicating whether to treat
columns as separate groups. Default is FALSE.}

\item{plot}{A logical value indicating whether to plot the control chart.
Default is TRUE.}

\item{side}{The side of the control chart,
can be "two.sided", "lower", or "upper".}
}
\value{
If `plot` is set to TRUE,
the function will generate a control chart plot.
If `plot` is set to FALSE,
the function will return the first out-of-control point (index)
based on the specified `side`.

If `plot` is set to TRUE,
the function will generate a control chart plot.
If `plot` is set to FALSE,
the function will return the first out-of-control point (index)
based on the specified `side`.
}
\description{
This function generates an EWMA (Exponentially Weighted Moving Average)
Control Chart on the Wilcoxon Rank Sum Test.
It compares a set of data points to a reference value and plots the control chart.

This function generates an EWMA (Exponentially Weighted Moving Average)
Control Chart on the Wilcoxon Rank Sum Test.
It compares a set of data points to a reference value and plots the control chart.
}
\examples{
# Generate EWMA Control Chart based on the Wilcoxon Rank Sum Test
X <- 10
Y <- matrix(rnorm(100), nrow = 10, ncol = 10)
ewma_wr(X, Y, lambda = 0.1, L = 3, plot = TRUE, side = 'two.sided')

# Generate EWMA Control Chart based on the Wilcoxon Rank Sum Test
X <- 10
Y <- matrix(rnorm(100), nrow = 10, ncol = 10)
ewma_wr(X, Y, lambda = 0.1, L = 3, plot = TRUE, side = 'two.sided')

}
