% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chart.R
\name{plot_chart}
\alias{plot_chart}
\title{Control Chart Plotting Function}
\usage{
plot_chart(ic, statistics, statistics2 = NULL, ucl, lcl, name = NULL, side)
}
\arguments{
\item{ic}{The in-control value for the control chart.}

\item{statistics}{A numeric vector of control statistics.}

\item{statistics2}{A numeric vector of additional control statistics (optional).}

\item{ucl}{The upper control limit value.}

\item{lcl}{The lower control limit value.}

\item{name}{The name of the control chart (optional).}

\item{side}{The side of the chart to plot. Must be one of 'two.sided', 'lower', or 'upper'.}
}
\description{
Plots the control chart based on the specified statistics and control limits.
}
\details{
The function plots the control chart based on the specified control statistics and control limits.
  It uses the `plot` function to create a line plot of the statistics.
  If additional statistics (`statistics2`) are provided, they are also plotted.
  The control limits are represented by dashed lines.
  The in-control value (`ic`) is represented by a dotted line.
  The points on the plot are colored red if they are outside the control limits, indicating an out-of-control condition.
}
