% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shewhart_mw.R
\name{shewhart_mw}
\alias{shewhart_mw}
\title{Shewhart Chart Based on Precedence Statistic}
\usage{
shewhart_mw(
  X,
  Y,
  cl,
  far = 0.0027,
  group_by_col = FALSE,
  plot = T,
  side = "two.sided"
)
}
\arguments{
\item{X}{The reference value.}

\item{Y}{The data matrix or vector.}

\item{far}{The false alarm rate (default: 0.0027).}

\item{group_by_col}{Logical value indicating whether to group the data by
columns (default: FALSE).}

\item{side}{The side of the control chart limits
("two.sided", "lower", or "upper") (default: "two.sided").}

\item{j}{The percentile value to calculate the
precedence statistic (default: round(nrow(Y)/2)).}
}
\value{
A character string indicating the used ARL0 (Average Run Length).
}
\description{
This function generates a Shewhart control chart based on the precedence
statistic for a given set of data.
}
\details{
The \code{shewhart_prec} function generates a Shewhart control
chart based on the precedence statistic.
The function calculates the precedence statistic for a given set of data
and constructs the control chart with
control limits based on the false alarm rate and the specified side.
It plots the control chart and returns
the used ARL0 (Average Run Length) as a character string.
}
\examples{
X <- rnorm(100)
Y <- matrix(rnorm(100), ncol = 5)
shewhart_mw(X, Y)

}
