% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shewhart_sn.R
\name{shewhart_sn}
\alias{shewhart_sn}
\title{Shewhart Control Chart Based on SN Statistic}
\usage{
shewhart_sn(X, med = 0, far = 0.0027, group_by_col = FALSE, side = "two.sided")
}
\arguments{
\item{X}{A matrix or data frame of observations.}

\item{med}{The center value for the control chart. Default is 0.}

\item{far}{The false alarm rate for control limits calculation. Default is 0.0027.}

\item{group_by_col}{Logical indicating whether the observations are grouped by column. 
If TRUE, the observations are transposed before further processing. Default is FALSE.}

\item{side}{The side of the control chart, can be "two.sided", "lower", or "upper".}
}
\value{
The Average Run Length (ARL) for the control chart.
}
\description{
This function constructs and plots a Shewhart Control Chart based on the SN statistic.
}
\examples{
# Create and plot a Shewhart Control Chart based on SN statistic
data <- matrix(rnorm(100), ncol = 2)
shewhart_sn(data, med = 0, far = 0.0027, group_by_col = FALSE, side = "two.sided")

}
