% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\docType{data}
\name{data2}
\alias{data2}
\title{EWMA-SN Data from Table 4.6 (Chakraborti and Graham, 2019)}
\format{
An example of a data matrix for the EWMA-SN
}
\usage{
data2
}
\description{
The data was extracted from the book of Chakraborti and Graham (2019)
that contains 10 groups with 15 observations each
}
\keyword{datasets}
