% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma_sn_arl_calc.R
\name{ewma_sn_arl_calc}
\alias{ewma_sn_arl_calc}
\title{Average Run Length (ARL) Calculation for EWMA Control Chart}
\usage{
ewma_sn_arl_calc(lambda = 0.1, L = 2.667, n = 1, sim = 10000)
}
\arguments{
\item{lambda}{The decay factor for the EWMA. Default is 0.1.}

\item{L}{The multiplier for the control limits. Default is 2.667.}

\item{n}{The number of observations in each simulation. Default is 1.}

\item{sim}{The number of simulations to perform. Default is 10000.}
}
\value{
A numeric vector representing the average run length (ARL) for each simulation.
}
\description{
Calculates the average run length (ARL) for the Exponentially Weighted Moving Average (EWMA) control chart.
}
\details{
The function calculates the average run length (ARL) for the EWMA control chart.
  It performs multiple simulations based on the specified parameters and returns the ARL for each simulation.
  The control limits are calculated based on the lambda and L parameters.
  The ARL is the number of observations until the control chart signals an out-of-control condition.
}
\examples{
# Calculate ARL for EWMA control chart
ewma_sn_arl_calc(lambda = 0.1, L = 2.667, n = 1, sim = 10000)

}
