% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shewhart_mw.R
\name{shewhart_mw}
\alias{shewhart_mw}
\title{Shewhart Control Chart based on Mann-Whitney Statistic}
\usage{
shewhart_mw(
  X,
  far = 0.0027,
  group_by_col = FALSE,
  phase1_num_samples = NULL,
  side = "two.sided"
)
}
\arguments{
\item{X}{A numeric matrix or data frame containing the data.}

\item{far}{The desired false alarm rate for the control chart. Default is 0.0027.}

\item{group_by_col}{A logical value indicating whether to treat columns as groups.
If TRUE, the matrix will be transposed. Default is FALSE.}

\item{phase1_num_samples}{The number of samples to include in phase 1. Default is NULL.}

\item{side}{The side of the chart to plot. Must be one of 'two.sided', 'lower', or 'upper'.
Default is 'two.sided'.}
}
\description{
Calculates and plots the Shewhart Control Chart based on the Mann-Whitney statistic.
}
\details{
The function calculates the Shewhart Control Chart based on the Mann-Whitney statistic.
  It divides the data into two phases: phase 1 and phase 2.
  Phase 1 consists of the specified number of samples, and phase 2 consists of the remaining samples.
  The Mann-Whitney statistic is computed between each sample in phase 2 and the phase 1 samples.
  The control limits are calculated based on the desired false alarm rate (far).
  The control chart is then plotted using the plot_chart function.
}
\examples{
# Generate example data
set.seed(42)
X <- matrix(rnorm(100), nrow = 20)

# Calculate Shewhart Control Chart based on Mann-Whitney statistic
shewhart_mw(X, far = 0.0027, group_by_col = FALSE, phase1_num_samples = 2, side = 'two.sided')

}
\seealso{
\code{\link{plot_chart}}
}
